fixed deleteing and editign todos

+++ server/controllers/Todos/deleteAll.js
@@ -2,7 +2,8 @@ import Todo from '../../models/todoModel'
 
 const deleteAllTodo = async (ctx) => {
   try {
-    const todos = await Todo.deleteMany()
+    const { _id: userId } = ctx.state.user
+    const todos = await Todo.deleteMany({ userId })
     console.log('all todos were deleted')
     ctx.resolve(todos)
   } catch (err) {
